Worked with JWT_Extended package

+++ resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from flask_restful import Resource, reqparse
 
 from models.item_model import ItemModel
@@ -34,7 +34,11 @@ class Item(Resource):
 
         return item.json(), 201  # 201 - Created | 202 - Accepted (when creation delay)
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required!'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -57,5 +61,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': list(map(lambda item: item.json(), ItemModel.query.all()))}
+        user_id = get_jwt_identity()
+        items = list(map(lambda item: item.json(), ItemModel.find_all()))
+        if user_id:
+            return {'items': items}
+        return {'items': [item['name'] for item in items],
+                'message': 'More data available if you log in.'}
